requirements.txt
python-docx>=1.1.0
pypdf>=4.0.0
python-pptx>=0.6.23
pymupdf>=1.24.0
pymupdf4llm>=0.0.15
pyinstaller>=6.11.0
